pyproject.toml
[tool.poetry]
name = "repl_python3_youtube-telegram-bot"
version = "0.1.0"
description = ""
authors = ["manodeveloper <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^13.9"
pytube = "^11.0.2"
replit = "^3.2.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
